refactor: rework business logic for VAI minting
